Moved config into the vcstools module
