+++ scripts/calibrate_vcs.py
@@ -30,7 +30,7 @@ from astropy.io import fits
 from vcstools.job_submit import submit_slurm
 from mdir import mdir
 from vcstools.metadb_utils import getmeta
-from config_vcs import load_config_file
+from vcstools.config import load_config_file
 
 logger = logging.getLogger(__name__)
