+++ scripts/cleanup.py
@@ -8,7 +8,7 @@ import re
 import logging
 logger = logging.getLogger(__name__)
 
-from config_vcs import load_config_file
+from vcstools.config import load_config_file
 
 def opt_parser():
     # Dictionary for choosing log-levels
