+++ vcstools/config.py
